requirements.txt
face_recognition
python-dotenv
pymongo
Pillow